refactor(security): move response middleware into module
