+++ app/src/server.rs
@@ -134,7 +134,9 @@ pub fn router(state: AppState) -> Router {
       state.clone(),
       maintenance_gate,
     ))
-    .layer(axum::middleware::from_fn(middlewares::security_headers))
+    .layer(axum::middleware::from_fn(
+      middlewares::security::security_headers,
+    ))
     .with_state(state)
 }
